Change the way connect() calls are JSONified to allow faceting.
The field "peer" used to contain "hostname ip:port", which makes
faceting hard because ES / Lucene will insist on tokenizing the
field prior to doing faceting.  So instead store only "ip:port"
in the "peer" field and add a new field, "host", that contains
the hostname of the peer to allow faceting.