Add Gearman support.
Through the --gearman (-G) flag, a slight change in behavior is
triggered so that the tracer can analyze what Gearman workers are
doing.  This a bit ad-hoc for now but can easily be generalized
to trace any client that communicates with server through read/write
syscalls on a given FD (which can be found by looking at the peer
after the `lsof' and on `connect' calls).